% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_mix.R
\name{compute_mix}
\alias{compute_mix}
\title{Compute the breathing gas mix for each dive segment and sets up the dive table for loadings computations}
\usage{
compute_mix(
  dive_segments,
  ppO2_low = 0.7,
  ppO2_high = 1.3,
  ppO2_switch_depth = 15,
  diluent = c(21, 0)
)
}
\arguments{
\item{dive_segments}{A tibble with dive segments as produced by the function \code{\link{create_dive_segments}}}

\item{ppO2_low}{Double. The low setting for the partial pressure of oxygen}

\item{ppO2_high}{Double. The high setting for the partial pressure of oxygen}

\item{ppO2_switch_depth}{Double. The depth at which the partial pressure of oxygen switches from the low to the high setting}

\item{diluent}{A numeric vector of length 2 with the percentage of oxygen, and helium in the diluent}
}
\value{
A tibble with the composition of the breathing gas at the start and end of each segment, the partial pressures of oxygen nitrogen and helium at the start and end of each segment, the equivalent air depth at the start and end of each segment, and columns for tissue loadings, tensions, M-values and percent gradients
}
\description{
Compute the breathing gas mix for each dive segment and sets up the dive table for loadings computations
}
\examples{
create_dive_segments(max_depth = 40, bottom_time = 50) |> 
compute_mix()
}
