% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_tissue_loadings.R
\name{compute_tissue_loadings}
\alias{compute_tissue_loadings}
\title{Compute compartment loadings for a dive table using the Buhlmann ZHL-16C model.}
\usage{
compute_tissue_loadings(dive_tbl, penalty = 3)
}
\arguments{
\item{dive_tbl}{Tibble. A dive table for which the loadings of the first segment have already been computed with the function \code{link{initialize_tissue_loadings}}.}

\item{penalty}{Integer. Either 1 (most permissive decompression model), 2 or 3 (default, most conservative decompression model).}
}
\value{
The same tibble as input, with the columns N2_load_start, N2_load_end, He_load_start and He_load_end filled with the tissue loadings for all segments of the dive.
}
\description{
Compute compartment loadings for a dive table using the Buhlmann ZHL-16C model.
}
\examples{
compute_tissue_loadings(dive_table)
}
