% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_stops_table.R
\name{create_stops_table}
\alias{create_stops_table}
\title{For a dive where decompression is done during true stops every 3 meters, computes the time for each deco stop up to the surface}
\usage{
create_stops_table(
  dive_tbl,
  gradient_low = 0.85,
  gradient_high = 0.9,
  first_stop,
  last_stop = 4,
  steps = 0.25,
  penalty = 3
)
}
\arguments{
\item{dive_tbl}{A tibble with each segment of the dive on a row}

\item{gradient_low}{Double. The low value of the factor gradient}

\item{gradient_high}{Double. The high value of the factor gradient}

\item{first_stop}{Double. The depth of the first stop in meters}

\item{last_stop}{Double. The depth of the last stop in meters}

\item{steps}{Double. The time increment (in decimal minutes) to add to each deco stop in order to reavch the appropriate stop duration. Smaller values give more precise results but will increase the computation time}

\item{penalty}{Integer. Either 1 (most permissive decompression model), 2 or 3 (default, most conservative decompression model).}
}
\value{
A tibble of dive segments limited to the decompression zone. Each segment is either a deco stop or an ascent segment between two stops
}
\description{
For a dive where decompression is done during true stops every 3 meters, computes the time for each deco stop up to the surface
}
\examples{
fs <- first_deco_stop(dive_table, gradient_low = 0.85)
create_stops_table(dive_table, first_stop = fs)
}
