% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deco_data.R
\name{deco_data}
\alias{deco_data}
\title{Compute M-values, tensions and percent_gradient for all segments}
\usage{
deco_data(dive_tbl)
}
\arguments{
\item{dive_tbl}{A tibble. Dive segments with loading already computed with \code{\link{compute_tissue_loadings}}}
}
\value{
The same tibble as input, with M-values, tensions and percent_gradient computed for all segments. Warning: at this stage, the values computed don't take into account any decompression stops. They will be used later to calculate the time needed for each deco stop.
}
\description{
Compute M-values, tensions and percent_gradient for all segments
}
\examples{
deco_data(dive_table)
}
