% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deco_last.R
\name{deco_last}
\alias{deco_last}
\title{Adjust the duration of the last stop so that the ascent from the last stop to the surface is done at a normal speed, and without exceeding the gradient_high value.}
\usage{
deco_last(dive_tbl, last_stop = 4, speed_asc = 10, penalty = 3, steps = 0.25)
}
\arguments{
\item{dive_tbl}{A tibble with the dive profile where each row is a segment of the dive. All sgemnts should be correct except for the last 2 (last stop and ascent from the last stop to the surface)}
}
\value{
The same tibble as \code{dive_tbl} but with the last 2 rows corrected
}
\description{
Adjust the duration of the last stop so that the ascent from the last stop to the surface is done at a normal speed, and without exceeding the gradient_high value.
}
\examples{
fst <- first_deco_stop(dive_table)
dive_table |> 
 deco_up_to_last(first_stop = fst) |> 
 deco_last()
 
}
