% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deco_up_to_last.R
\name{deco_up_to_last}
\alias{deco_up_to_last}
\title{Computes the time for each deco stop up to the last one (its duration will have to be adjusted later)}
\usage{
deco_up_to_last(
  dive_tbl,
  gradient_low = 0.85,
  gradient_high = 0.9,
  first_stop,
  last_stop = 4,
  steps = 0.25,
  penalty = 3
)
}
\arguments{
\item{dive_tbl}{A tibble with each segment of the dive on a row}

\item{gradient_low}{Double. The low value of the factor gradient}

\item{gradient_high}{Double. The high value of the factor gradient}

\item{first_stop}{Double. The depth of the first stop in meters}

\item{last_stop}{Double. The depth of the last stop in meters}

\item{steps}{Double. The time increment (in decimal minutes) to add to each deco stop in order to reavch the appropriate stop duration. Smaller values give more precise results but will increase the computation time}

\item{penalty}{Integer. Either 1 (most permissive decompression model), 2 or 3 (default, most conservative decompression model).}
}
\value{
The same tibble as \code{dive_tbl} but with \code{phase}, \code{time_end} and \code{duration} adjusted for each deco stop. All loadings, tensions, M-values and percent gradient are also re-computed with the correct deco stop durations.
}
\description{
Computes the time for each deco stop up to the last one (its duration will have to be adjusted later)
}
