% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deco_up_to_next.R
\name{deco_up_to_next}
\alias{deco_up_to_next}
\title{Computes the time for one deco stop and the ascent up to the next stop. Necessary when the decompression is done with true deco stops every 3 meters.}
\usage{
deco_up_to_next(dive_tbl, steps = 0.25, penalty = 3)
}
\arguments{
\item{dive_tbl}{A tibble with only segments for two consecutive stops and ascent segments between them}

\item{steps}{Double. The time increment (in decimal minutes) to add to each deco stop in order to reavch the appropriate stop duration. Smaller values give more precise results but will increase the computation time}

\item{penalty}{Integer. Either 1 (most permissive decompression model), 2 or 3 (default, most conservative decompression model).}
}
\value{
The same tibble as \code{dive_tbl} but with \code{phase}, \code{time_end} and \code{duration} adjusted for each deco stop. All loadings, tensions, M-values and percent gradient are also re-computed with the correct deco stop durations.
}
\description{
Computes the time for one deco stop and the ascent up to the next stop. Necessary when the decompression is done with true deco stops every 3 meters.
}
\details{
This function is used when the decompression is not just a slow and continuous ascent, but when true stops are made every 3 meters from the first stop to the surface. The duration of a stop is set so that the ascent to the next stop is possible without going above the factor gradient set for the next stop.
}
