% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deco_zone.R
\name{deco_zone}
\alias{deco_zone}
\title{Determines the depth at which the diver enters the decompression zone}
\usage{
deco_zone(dive_tbl)
}
\arguments{
\item{dive_tbl}{A Tibble containing information for all segments of a dive}
}
\value{
A numeric (\code{<int>}) value indicating the depth at which the diver enters the decompression zone
}
\description{
Determines the depth at which the diver enters the decompression zone
}
\examples{
deco_zone(dive_table)
}
