% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DeepCCR-package.R
\docType{package}
\name{DeepCCR-package}
\alias{DeepCCR}
\alias{DeepCCR-package}
\title{DeepCCR: Deco Planning of CCR Deep Dives}
\description{
Given a set of parameters for a dive (e.g. maximum depth, time, gas mixes available, gradients low and high, descent and ascent rates...) DeepCCR uses the factor-gradient Bühlmann model with 16 compartments (ZHL-16c) to provide the full decompression planning for the dive with a closed circuit rebreather.
}
\keyword{internal}
