% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dive_profile.R
\name{dive_profile}
\alias{dive_profile}
\title{Given a set of parameters for a dive, \code{profile} calculates the dive profile by slicing the dive into small segments of 1 meters and computing all decompression data for each segment up to the surface.}
\usage{
dive_profile(
  max_depth,
  bottom_time,
  speed_desc,
  speed_asc,
  last_stop,
  ppO2_low,
  ppO2_high,
  ppO2_switch_depth,
  diluent,
  penalty,
  gradient_low,
  gradient_high,
  steps,
  deco_type = c("stops", "continuous")
)
}
\arguments{
\item{max_depth}{Double. The maximum depth of the dive (in meters)}

\item{bottom_time}{Double. Total dive time (in minutes) befor starting the ascent}

\item{speed_desc}{Double. The speed of the descent (in meters per minute)}

\item{speed_asc}{Double. The speed of the ascent (in meters per minute)}

\item{last_stop}{Double. The depth of the last decompression stop (in meters)}

\item{ppO2_low}{Double. The low setting for the partial pressure of oxygen}

\item{ppO2_high}{Double. The high setting for the partial pressure of oxygen}

\item{ppO2_switch_depth}{Double. The depth at which the partial pressure of oxygen switches from the low to the high setting}

\item{diluent}{A numeric vector of length 2 with the percentage of oxygen, and helium in the diluent}

\item{penalty}{Integer. Either 1 (most permissive decompression model), 2 or 3 (default, most conservative decompression model)}

\item{gradient_low}{Double. The low value of the factor gradient}

\item{gradient_high}{Double. The high value of the factor gradient}

\item{steps}{Double. The time increment (in decimal minutes) to add to each deco stop in order to reavch the appropriate stop duration. Smaller values give more precise results but will increase the computation time}

\item{deco_type}{Character. Either "stops" (default) or "continuous". If "stops", the ascent is done by making stops every 3 meters from the first stop to the last stop. If "continuous", the decompression is done continuously, as a slow ascent from the first stop to the last stop.}
}
\value{
A tibble with the dive profile where each row is a segment of the dive. See \code{dive_table} for a detailed description of each variable
}
\description{
Given a set of parameters for a dive, \code{profile} calculates the dive profile by slicing the dive into small segments of 1 meters and computing all decompression data for each segment up to the surface.
}
\examples{
dive <- dive_profile(max_depth = 40, bottom_time = 50, speed_desc = 20, speed_asc = 10, 
        last_stop = 4, ppO2_low = 0.7, ppO2_high = 1.3, 
        ppO2_switch_depth = 15, diluent = c(21, 00), penalty = 3, 
        gradient_low = 0.85, gradient_high = 0.90, steps = 0.25, 
        deco_type = "stops")
dive |> print(n = Inf)
}
