% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EAD.R
\name{EAD}
\alias{EAD}
\title{Computes the Equivalent Air Depth (EAD) of one or several gas at any depth}
\usage{
EAD(N2, depth)
}
\arguments{
\item{N2}{Double. Fraction of N2 in the gas mix}

\item{depth}{Double. Depth of the dive in meters}
}
\value{
Double. Equivalent Air Depth (EAD) in meters
}
\description{
Computes the Equivalent Air Depth (EAD) of one or several gas at any depth
}
\examples{
EAD(N2 = 0.60, depth = 30)
EAD(N2 = 0.30, depth = 100)
}
