% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/first_deco_stop.R
\name{first_deco_stop}
\alias{first_deco_stop}
\title{Determines the Depth of the First Deco Stop}
\usage{
first_deco_stop(dive_tbl, gradient_low = 0.85)
}
\arguments{
\item{dive_tbl}{A tibble with one row for each segment of the dive}

\item{gradient_low}{Double. The minimum gradient for a stop to be considered a deco stop}
}
\value{
Double. The depth (in meters) of the first deco stop
}
\description{
Determines the Depth of the First Deco Stop
}
\examples{
first_deco_stop(dive_table, gradient_low = 0.85)
first_deco_stop(dive_table, gradient_low = 0.30)
}
