% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GF_stops.R
\name{GF_stops}
\alias{GF_stops}
\title{I don't know what this function does, but it does it good...}
\usage{
GF_stops(a, b, ppart, GF, prof.max)
}
\arguments{
\item{a}{numeric vector of Bühlmann coefficients (length 16)}

\item{b}{numeric vector of Bühlmann coefficients (length 16)}

\item{ppart}{numeric vector of partial pressures (length 2)}

\item{GF}{numeric value of gradient factor}

\item{prof.max}{Double. Maximum depth}
}
\description{
I don't know what this function does, but it does it good...
}
