% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/increase_stop_duration.R
\name{increase_stop_duration}
\alias{increase_stop_duration}
\title{Increases the duration of the last stop}
\usage{
increase_stop_duration(last_tbl, steps, penalty = 3)
}
\arguments{
\item{last_tbl}{A data frame with the last 2 segment of the dive}

\item{steps}{Double. How much time should we add to the stop (decimal minutes)}

\item{penalty}{Integer. The penalty for the tissue loadings}
}
\description{
Increases the duration of the last stop
}
