% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialize_tissue_loadings.R
\name{initialize_tissue_loadings}
\alias{initialize_tissue_loadings}
\title{Computes tissue loadings in N2 and He for the first segment of a dive}
\usage{
initialize_tissue_loadings(dive_tbl, penalty = 3)
}
\arguments{
\item{dive_tbl}{A tibble with dive segments as produced by \code{\link{compute_mix}}}

\item{penalty}{Integer. Penalty value to be used in the calculation of the tissue loadings. Possible values are 1, 2 and 3 (default, the most conservative model)}
}
\value{
The same tibble as input, with the columns N2_load_start, N2_load_end, He_load_start and He_load_end filled with the tissue loadings for the first segment of the dive
}
\description{
Computes tissue loadings in N2 and He for the first segment of a dive
}
