% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loop_mix.R
\name{loop_mix}
\alias{loop_mix}
\title{Compute the mix of gases in the loop to achieve a desired ppO2 at a given depth}
\usage{
loop_mix(diluent, ppO2, depth)
}
\arguments{
\item{diluent}{A numeric vector of length 3 with the fraction of O2, N2, and He in the diluent}

\item{ppO2}{Double. The desired ppO2 at the depth}

\item{depth}{Double. The depth in meters}
}
\value{
A numeric vector of length 3 with the fraction of O2, N2, and He in the mix
}
\description{
Compute the mix of gases in the loop to achieve a desired ppO2 at a given depth
}
\examples{
loop_mix(c(0.21, 0), 1.3, 45)
}
