% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/percent_gradient.R
\name{percent_gradient}
\alias{percent_gradient}
\title{Given the M-values, depth, and inert gas partial pressures in each compartment, this function calculates the percentage of the gradient achieved through saturation. In other words, it determines how close the compartment tensions are to the M-values, expressed as a percentage of the gradient from ambient pressure.}
\usage{
percent_gradient(tension, depth, M_val)
}
\arguments{
\item{tension}{numeric. The tension in the compartment (sum of the ppart of each gas)}

\item{depth}{numeric. The depth in meters}

\item{M_val}{numeric vector of M-values (16 values)}
}
\description{
Given the M-values, depth, and inert gas partial pressures in each compartment, this function calculates the percentage of the gradient achieved through saturation. In other words, it determines how close the compartment tensions are to the M-values, expressed as a percentage of the gradient from ambient pressure.
}
