% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Tissue_loadings.R
\name{Tissue_loadings}
\alias{Tissue_loadings}
\title{Computes loadings in N2 and He for the 16 tissues defined by Bühlmann (table ZH-L16 C)}
\usage{
Tissue_loadings(Segment, Penalty = 3)
}
\arguments{
\item{Segment}{a one row tibble containing (at least) the following variables: N2, He, Start_depth, End_depth, Duration, N2_Load_Start and He_Load_Start}

\item{Penalty}{either 1 (most permissive decompression model), 2 or 3 (default, most conservative decompression model)}
}
\description{
Computes loadings in N2 and He for the 16 tissues defined by Bühlmann (table ZH-L16 C)
}
