% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Create_ZHL16_C.R
\name{ZHL16_C}
\alias{ZHL16_C}
\title{Creates a tibble containing each of Bühlmann's values for its 16-compartment model}
\description{
Creates a tibble containing each of Bühlmann's values for its 16-compartment model
}
