% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_dive_segments.R
\name{create_dive_segments}
\alias{create_dive_segments}
\title{Creates a table with all segments of a dive}
\usage{
create_dive_segments(
  max_depth,
  bottom_time,
  speed_desc = 20,
  speed_asc = 10,
  last_stop = 4
)
}
\arguments{
\item{max_depth}{Double. The maximum depth of the dive (in meters)}

\item{bottom_time}{Double. Total dive time (in minutes) befor starting the ascent}

\item{speed_desc}{Double. The speed of the descent (in meters per minute)}

\item{speed_asc}{Double. The speed of the ascent (in meters per minute)}

\item{last_stop}{Double. The depth of the last decompression stop (in meters)}
}
\value{
A 5-column tibble with the diving phase, the start and end depth and time for each segment of the dive
}
\description{
Creates a table with all segments of a dive
}
\examples{
create_dive_segments(max_depth = 40, bottom_time = 50)
}
