% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EAD.R
\name{EAD}
\alias{EAD}
\title{EAD computes the Maximum Equivalent Air Depth (EAD) of one or several gas at Max_depth
N2: tibble, vector or scalar (dbl). Fraction of N2 in the gas mix
depth: tibble, vector or scalar (dlb). Current depth of the dive in meters (msw)}
\usage{
EAD(N2, depth)
}
\description{
EAD computes the Maximum Equivalent Air Depth (EAD) of one or several gas at Max_depth
N2: tibble, vector or scalar (dbl). Fraction of N2 in the gas mix
depth: tibble, vector or scalar (dlb). Current depth of the dive in meters (msw)
}
