% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/last_ascent.R
\name{last_ascent}
\alias{last_ascent}
\title{Recomputes the loadings of the last ascent segment}
\usage{
last_ascent(last_tbl, penalty = 3)
}
\arguments{
\item{last_tbl}{A data frame containing the last ascent segment}
}
\description{
Recomputes the loadings of the last ascent segment
}
