% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/M_value.R
\name{M_value}
\alias{M_value}
\title{Computes the M-values at the end of a segment, for each of the 16 tissues defined by Bühlmann (table ZH-L16 C), for both N2 and He.}
\usage{
M_value(N2_Load_End, He_Load_End, End_depth)
}
\arguments{
\item{N2_Load_End}{Double. The N2 loading at the end of the segment.}

\item{He_Load_End}{Double. The He loading at the end of the segment.}

\item{End_depth}{Double. The depth at the end of the segment.}
}
\description{
Computes the M-values at the end of a segment, for each of the 16 tissues defined by Bühlmann (table ZH-L16 C), for both N2 and He.
}
