% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/M_value.R
\name{M_value}
\alias{M_value}
\title{Computes the N2 and He M-values at the end of a segment for each of the 16 compartments defined by Bühlmann's ZH-L16 C model.}
\usage{
M_value(N2_load_end, He_load_end, depth_end)
}
\arguments{
\item{N2_load_end}{Numeric vector. The N2 loadings of the 16 compartments at the end of the segment.}

\item{He_load_end}{Numeric vector. The He loadings of the 16 compartments at the end of the segment.}

\item{depth_end}{Double. The depth at the end of the segment.}
}
\value{
A numeric vector with the M-values for each of the 16 compartments at the end of the segment.
}
\description{
Computes the N2 and He M-values at the end of a segment for each of the 16 compartments defined by Bühlmann's ZH-L16 C model.
}
\examples{
bottom <- dive_table |>
 dplyr::filter(phase == "bottom")
M_value(N2_load_end = bottom$N2_load_end, 
        He_load_end = bottom$He_load_end, 
        depth_end = bottom$depth_end)
        
}
