% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tissue_loadings.R
\name{tissue_loadings}
\alias{tissue_loadings}
\title{Computes loadings in N2 and He for the 16 tissues defined by Bühlmann (table ZH-L16 C)}
\usage{
tissue_loadings(
  N2,
  He,
  depth_start,
  depth_end,
  duration,
  N2_load_start,
  He_load_start,
  penalty = 3
)
}
\arguments{
\item{N2}{Double. The fraction of N2 in the gas mix}

\item{He}{Double. The fraction of He in the gas mix}

\item{depth_start}{Double. The depth at the start of the segment in meters}

\item{depth_end}{Double. The depth at the end of the segment in meters}

\item{duration}{Double. The duration of the segment in minutes (or in fraction of a minute)}

\item{N2_load_start}{a vector of 16 values representing the N2 loadings at the start of the segment}

\item{He_load_start}{a vector of 16 values representing the He loadings at the start of the segment}

\item{penalty}{either 1 (most permissive decompression model), 2 or 3 (default, most conservative decompression model)}
}
\description{
Computes loadings in N2 and He for the 16 tissues defined by Bühlmann (table ZH-L16 C)
}
